app/requirements.txt
numpy
pandas
Flask
scikit-learn
torch
torchvision
requests
Pillow
gunicorn
authlib
python-dotenv
pymongo
markupsafe
typing
pickle-mixin
flask-cors
plotly
seaborn
statsmodels
